case-study-app/requirements.txt
Flask
gunicorn==20.1.0; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
google-cloud-datastore==2.15.2
google-cloud-language==2.9.1
flask-restx==1.1.0
scikit-learn==1.2.2
werkzeug==2.3.4
beautifulsoup4==4.12.2
dash
dash-bootstrap-components==1.4.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
gensim==4.3.1
google-api-core==2.11.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-pasta==0.2.0
googleapis-common-protos==1.59.1
langcodes==3.3.0
nbformat==5.9.1
nltk==3.8.1
pandas==2.0.3
plotly==5.15.0
regex==2023.6.3
scipy==1.10.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4